requirements.txt
Django==4.2.15
psycopg2==2.9.9
django-braces==1.14.0
django-model-utils==4.0.0
django-role-permissions==2.2.1

raven==6.10.0
django-sentry==1.13.5
requests>=2.23.0
pytz
# support special characters

django-allauth==0.63.3

# Django Easy Audit
# https://github.com/soynatan/django-easy-audit
django-easy-audit==v1.3.7.b3
django-grappelli==3.0.7

djangorestframework==3.15.2
djangorestframework-gis==1.0
django-colorfield==0.9.0
pymemcache==4.0.0
# bibliography
bibtexparser==1.1.0
eutils==0.6.0
habanero==0.7.2

# documentation generation
django-modelsdoc==0.1.11

# contacts page
django-contact-us==0.4.3
django-ordered-model==3.7.4

# Statistics Dashboard

# prometheus monitoring support
django-prometheus==2.0.0
prometheus_client==0.7.1

pyshp==2.1.0

# for django flatpages
django-ckeditor==6.6.1
django-crispy-forms==1.9.1
# git+https://github.com/dimasciput/pandas_access.git

selenium==3.141.0
django-admin-rangefilter==0.6.0
django-preferences==1.0.0
PyGithub==1.51
sorl-thumbnail==12.9.0
svglib==1.0.0
simplejson==3.17.0
django-extensions==3.2.3
service-identity==18.1.0
geopy==2.0.0
python-dwca-reader==0.16.0
django-json-widget==1.1.1
django-webpack-loader==1.3.0
requests-cache==0.6.0
dj-database-url==0.5.0

django-polymorphic==3.1.0
django-modeltranslation==0.16.2
django-autocomplete-light==3.9.7
dj-pagination==2.5.0
django-mptt==0.14.0
django-guardian==2.3.0
djangorestframework-guardian==0.3.0
django-cors-headers==4.3.1
django-invitations==2.0.0
django-uuid-upload-path==1.0.0
django-pipeline==2.1.0
python-memcached==1.59
geonode-oauth-toolkit==2.1.0
mock==4.0.3
django-imagekit==4.0.2
pika==1.2.0

MarkupSafe==1.1.1
pygbif==0.6.0

pycurl==7.45.2
django-forms-bootstrap==3.1.0
importlib-metadata==4.11.4
django-taggit==5.0.1
django-treebeard==4.7
django-filter==23.5
sentry-sdk==1.40.0

django-admin-inline-paginator==0.4.0
pandas==2.2.1
openpyxl==3.1.2
celery==5.3.6
django-celery-results==2.5.1
django-tenants==3.6.1
tenant-schemas-celery==2.2.0

# Library to handle geospatial functions
git+https://github.com/kartoza/CloudNativeGIS.git@bims
GeoAlchemy2==0.15.1
geopandas==0.14.4

drf-nested-routers==0.93.5
drf-yasg==1.21.7

